rename,update dependencies
