+++ build.gradle
@@ -19,7 +19,6 @@ apply plugin: 'java'
 apply plugin: 'eclipse'
 apply plugin: 'idea'
 apply plugin: 'org.springframework.boot'
-apply plugin: 'com.palantir.docker'
 
 sourceCompatibility = 1.8
