+++ gedit-cloud-hello-graphql/src/main/resources/application.properties
@@ -2,7 +2,7 @@ org.springframework.boot.autoconfigure.EnableAutoConfiguration=org.ditto.easyhan
 security.basic.enabled=false
 management.security.enabled=false
 grpc.port=8980
-server.port=8088
+server.port=9000
 gedit.docker.enabled=false
 gedit.cloud.config.server.git.uri=
 gedit.cloud.config.server.git.clone-on-start=true
