+++ gedit-cloud-hellogrpc1/src/main/java/com/github/conanchen/gedit/hello/Grpc1HelloApplication.java
@@ -1,7 +1,7 @@
 package com.github.conanchen.gedit.hello;
 
 
-import com.github.conanchen.gedit.hello.controller.HelloGrpcClient;
+import com.github.conanchen.gedit.hello.client.HelloGrpc2Client;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.Bean;
@@ -11,10 +11,5 @@ public class Grpc1HelloApplication {
 
     public static void main(String[] args) {
         SpringApplication.run(Grpc1HelloApplication.class, args);
-    }
-
-    @Bean
-    public HelloGrpcClient helloGrpcClient() {
-        return new HelloGrpcClient();
     }
 }
