+++ gedit-cloud-hellogrpc1/src/main/resources/application.properties
@@ -2,6 +2,8 @@ org.springframework.boot.autoconfigure.EnableAutoConfiguration=org.ditto.easyhan
 security.basic.enabled=false
 management.security.enabled=false
 grpc.port=8901
+server.port=9001
+hellogrpc2.port=8902
 gedit.docker.enabled=false
 gedit.cloud.config.server.git.uri=
 gedit.cloud.config.server.git.clone-on-start=true
