+++ gedit-cloud-hellogrpc2/src/main/resources/application.properties
@@ -1,7 +1,8 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=org.ditto.easyhan.config.SpringConfig
 security.basic.enabled=false
 management.security.enabled=false
-grpc.port=8900
+grpc.port=8902
+server.port=9002
 gedit.docker.enabled=false
 gedit.cloud.config.server.git.uri=
 gedit.cloud.config.server.git.clone-on-start=true
