--- src/main/resources/graphql/blog.graphqls
@@ -1,33 +1,0 @@
-type Post {
-    id: ID!
-    title: String!
-    text: String!
-    category: String
-    author: Author
-}
-
-type Author {
-    id: ID!
-    name: String!
-    thumbnail: String
-    posts: [Post]!
-}
-
-
-input PostInput {
-    title: String!
-    text: String!
-    category: String
-    authorId: String!
-}
-
-# The Root Query for the application
-type Query {
-    recentPosts(count: Int, offset: Int): [Post]!
-}
-
-# The Root Mutation for the application
-type Mutation {
-    writePost(title: String!, text: String!, category: String, author: String!) : Post!
-    writePostWithInput(input: PostInput!) : Post!
-}
